Separate HDR upload and JSONL upload components out, only call MKDIR for the local file system, improve naming to support both types of uploads.
